uiextra: uitech: Change tech bonus indicator and clickable area
In 1oom v1.0 there is a small rectangle crammed in between the tech level and edge
indicating whether max tech bonus received for that research field.
This indicator looks bad, having a slight colour change on the tech level number
appears more pleasant for the eye.
---
Clicking on the tech level will adjust the research sliders so that it is maximum
bonus level.

Updated (MyName):
Combined 2 commits, removed unnecessary comment.