hwx11: Add an X11 "hw" interface for the classical MOO UI
This is another graphical interface alongside the SDL and Allegro ones.
The main difference is that X11 is more standard, has less dependencies,
and does not grab the mouse or alter screen resolutions or settings.
Also the implementation is shorter and more straightforward than the
other interfaces.

The drawback with X11 is that it does not do audio, instead future work
might include using ALSA in conjunction with X for a full feature client.

Surprisingly, the X11 graphics appear much faster than the SDL interfaces,
with the result that some animations (battle movement) might need delays.