hwsdl1: Add support for detecting ctrl keys
By default a press of ctrl key is reported as a key (left or right ctrl)
without the ctrl modifier being indicated.
The result was that holding down ctrl key was not detectable by reading
the 1oom keyboard modifier mask.

This patch reports a control key as "Ctrl + Control_L" instead of plain
"Control_L" (or "Ctrl + Control_R" instead of "Control_R")