audience: Correctly fill in broken treaty placeholder
The placeholder for broken treaties by an AI was filled in with the last
treaty broken by the player with this AI, defaulting to trade agreement
if no treaty was broken. It now fills in the correct treaty.

I added a few comments to explain some of the placeholders while at it.

Fixes #18