game_planet.c: Don't use negative terraforming cost
When conquering a planet from a race with more advanced terraforming
technology, the game would calculate a negative terraforming cost when
predicting the results of ECO spending. This would then lead to vastly
inflated population growth estimates. Actual population growth was not
affected by this.

This behavior was present in MOO 1.3, but it obviously is a bug, so it
is fixed here.

While at it, don't reuse the "adjusted waste" variable for calculating
terraforming costs. Let's leave micro-optimizations like that to the
compiler.

Fixes (parts of) #21