implemented the Rules options (pbx, fix bugs, MOO v1.3), see doc/usage_classic.txt
It's a messy way to deal with a messy situation, but we need a way to make the gameplay bugfixes more accesible than having to compile and then apply fixbugs.pbx.
Also, this commit activates the game_num_monster_rest_att fix in pbxin_fixbugs.txt and creates game_num_soil_rounding_fix (only the variable, not the actual effect) so that it can be appiled independently from the uifeatures patch.