Infer nebula types when loading a save
The game does not store the nebula types in the save file which means
that after loading a game, they were displayed incorrectly. The nebula
type can be inferred from the bounding boxes, though, so we do that as
a quick fix until we move to a new save format.

Based on a merge request by Ignatius:
https://gitlab.com/Tapani_/1oom/merge_requests/10