Fixed embarrassing bug with new rules option
The rules take effect in game_aux_start(), however, tbl_startship is
copied to the design table during game_generate_home_etc() which is
executed earlier. So the ship cost bug is not fixed by the "fix bug"
ruleset. The culprit was a last minute optimization (never a good
idea) to save one call of game_aux_set_rules().