save format, uinewgame, uigameopts: add game options support
This commits uses 24 bytes unused from the savefile header to
allow for up to 23 game options. This replaces the xoptions
byte. The new options are available in the New Game screen
under the Rules submenu and in the Game options screen.

opt.retreat replaces game_num_retreat_redir_fix which is
ignored but left as not to break pbx-patches. The RBO option
is as stub and currently unimplemented.