Reapply 45aebab5b54e90b0706aceadce18912141a26084
The seen structure is used only by rendering functions.
Classic AI does not use this structure, so in the original it is
not filled in for AI. Filling out this structure does not change
the behavior, but allows it to be used in new AIs.
This idea was implemented in 1oom v0.1